parallel magics are no longer an extension
